+++ cedar-policy-core/src/entities/json/err.rs
@@ -17,9 +17,8 @@
 use std::fmt::Display;
 
 use super::SchemaType;
-use crate::ast::{
-    EntityType, EntityUID, Expr, ExprKind, Name, PolicyID, RestrictedExpr, RestrictedExprError,
-};
+use crate::ast::{EntityUID, Expr, ExprKind, Name, PolicyID, RestrictedExpr, RestrictedExprError};
+use crate::entities::conformance::{EntitySchemaConformanceError, HeterogeneousSetError};
 use crate::extensions::ExtensionFunctionLookupError;
 use crate::parser::err::ParseErrors;
 use either::Either;
@@ -64,9 +63,6 @@ pub enum JsonDeserializationError {
     /// Restricted expression error
     #[error(transparent)]
     RestrictedExpressionError(#[from] RestrictedExprError),
-    /// An error occurred when looking up an extension function
-    #[error(transparent)]
-    FailedExtensionFunctionLookup(#[from] ExtensionFunctionLookupError),
     /// A field that needs to be a literal entity reference, was some other JSON value
     #[error("{ctx}, expected a literal entity reference, but got `{}`", display_json_value(.got.as_ref()))]
     ExpectedLiteralEntityRef {
@@ -108,45 +104,21 @@ pub enum JsonDeserializationError {
         /// argument type of the constructor we were looking for
         arg_type: Box<SchemaType>,
     },
-    /// During schema-based parsing, encountered an entity of a type which is
-    /// not declared in the schema. Note that this error is only used for non-Action entity types.
-    #[error("entity `{uid}` has type `{}` which is not declared in the schema{}",
-        &.uid.entity_type(),
-        match .suggested_types.as_slice() {
-            [] => String::new(),
-            [ty] => format!(". Did you mean `{ty}`?"),
-            tys => format!(". Did you mean one of {:?}?", tys.iter().map(ToString::to_string).collect::<Vec<String>>())
-        }
-    )]
-    UnexpectedEntityType {
-        /// Entity that had the unexpected type
-        uid: EntityUID,
-        /// Suggested similar entity types that actually are declared in the schema (if any)
-        suggested_types: Vec<EntityType>,
+    /// The same key appears two or more times in a single record literal
+    #[error("{ctx}, duplicate key `{key}` in record literal")]
+    DuplicateKeyInRecordLiteral {
+        /// Context of this error
+        ctx: Box<JsonDeserializationErrorContext>,
+        /// The key that appeared two or more times
+        key: SmolStr,
     },
-    /// During schema-based parsing, encountered an action which was not
-    /// declared in the schema
-    #[error("found action entity `{uid}`, but it was not declared as an action in the schema")]
-    UndeclaredAction {
-        /// Action which was not declared in the schema
-        uid: EntityUID,
-    },
-    /// During schema-based parsing, encountered an action whose definition
-    /// doesn't precisely match the schema's declaration of that action
-    #[error("definition of action `{uid}` does not match its schema declaration")]
-    ActionDeclarationMismatch {
-        /// Action whose definition mismatched between entity data and schema
-        uid: EntityUID,
-    },
-    /// During schema-based parsing, encountered this attribute on this entity, but that
-    /// attribute shouldn't exist on entities of this type
-    #[error("attribute `{attr}` on `{uid}` should not exist according to the schema")]
-    UnexpectedEntityAttr {
-        /// Entity that had the unexpected attribute
-        uid: EntityUID,
-        /// Name of the attribute that was unexpected
-        attr: SmolStr,
-    },
+    /// During schema-based parsing, encountered an entity which does not
+    /// conform to the schema.
+    ///
+    /// This error contains the `Entity` analogues some of the other errors
+    /// listed below, among other things.
+    #[error(transparent)]
+    EntitySchemaConformance(EntitySchemaConformanceError),
     /// During schema-based parsing, encountered this attribute on a record, but
     /// that attribute shouldn't exist on that record
     #[error("{ctx}, record attribute `{record_attr}` should not exist according to the schema")]
@@ -155,15 +127,6 @@ pub enum JsonDeserializationError {
         ctx: Box<JsonDeserializationErrorContext>,
         /// Name of the (Record) attribute which was unexpected
         record_attr: SmolStr,
-    },
-    /// During schema-based parsing, didn't encounter this attribute of an
-    /// entity, but that attribute should have existed
-    #[error("expected entity `{uid}` to have an attribute `{attr}`, but it does not")]
-    MissingRequiredEntityAttr {
-        /// Entity that is missing a required attribute
-        uid: EntityUID,
-        /// Name of the attribute which was expected
-        attr: SmolStr,
     },
     /// During schema-based parsing, didn't encounter this attribute of a
     /// record, but that attribute should have existed
@@ -174,48 +137,58 @@ pub enum JsonDeserializationError {
         /// Name of the (Record) attribute which was expected
         record_attr: SmolStr,
     },
-    /// During schema-based parsing, the given attribute on the given entity had
-    /// a different type than the schema indicated to expect
-    #[error("{ctx}, type mismatch: attribute was expected to have type {expected}, but actually has type {actual}")]
+    /// During schema-based parsing, found a different type than the schema indicated.
+    ///
+    /// (This is used in all cases except inside entity attributes; type mismatches in
+    /// entity attributes are reported as `Self::EntitySchemaConformance`. As of
+    /// this writing, that means this should only be used for schema-based
+    /// parsing of the `Context`.)
+    #[error("{ctx}, type mismatch: expected type {expected}, but actually has type {actual}")]
     TypeMismatch {
-        /// Context of this error
+        /// Context of this error, which will be something other than `EntityAttribute`.
+        /// (Type mismatches in entity attributes are reported as
+        /// `Self::EntitySchemaConformance`.)
         ctx: Box<JsonDeserializationErrorContext>,
         /// Type which was expected
         expected: Box<SchemaType>,
         /// Type which was encountered instead
         actual: Box<SchemaType>,
     },
-    /// During schema-based parsing, found a set whose elements don't all have the
-    /// same type.  This doesn't match any possible schema.
-    #[error("{ctx}, set elements have different types: {ty1} and {ty2}")]
+    /// During schema-based parsing, found a set whose elements don't all have
+    /// the same type.  This doesn't match any possible schema.
+    ///
+    /// (This is used in all cases except inside entity attributes;
+    /// heterogeneous sets in entity attributes are reported as
+    /// `Self::EntitySchemaConformance`. As of this writing, that means this
+    /// should only be used for schema-based parsing of the `Context`. Note that
+    /// for non-schema-based parsing, heterogeneous sets are not an error.)
+    #[error("{ctx}, {err}")]
     HeterogeneousSet {
-        /// Context of this error
+        /// Context of this error, which will be something other than `EntityAttribute`.
+        /// (Heterogeneous sets in entity attributes are reported as
+        /// `Self::EntitySchemaConformance`.)
         ctx: Box<JsonDeserializationErrorContext>,
-        /// First element type which was found
-        ty1: Box<SchemaType>,
-        /// Second element type which was found
-        ty2: Box<SchemaType>,
+        /// Underlying error
+        err: HeterogeneousSetError,
     },
-    /// The same key appears two or more times in a single record literal
-    #[error("{ctx}, duplicate key `{key}` in record literal")]
-    DuplicateKeyInRecordLiteral {
-        /// Context of this error
+    /// During schema-based parsing, error looking up an extension function.
+    /// This error can occur during schema-based parsing because that may
+    /// require getting information about any extension functions referenced in
+    /// the JSON.
+    ///
+    /// (This is used in all cases except inside entity attributes; extension
+    /// function lookup errors in entity attributes are reported as
+    /// `Self::EntitySchemaConformance`. As of this writing, that means this
+    /// should only be used for schema-based parsing of the `Context`.)
+    #[error("{ctx}, {err}")]
+    ExtensionFunctionLookup {
+        /// Context of this error, which will be something other than
+        /// `EntityAttribute`.
+        /// (Extension function lookup errors in entity attributes are reported
+        /// as `Self::EntitySchemaConformance`.)
         ctx: Box<JsonDeserializationErrorContext>,
-        /// The key that appeared two or more times
-        key: SmolStr,
-    },
-    /// During schema-based parsing, found a parent of a type that's not allowed
-    /// for that entity
-    #[error(
-        "{ctx}, `{uid}` is not allowed to have a parent of type `{parent_ty}` according to the schema"
-    )]
-    InvalidParentType {
-        /// Context of this error
-        ctx: Box<JsonDeserializationErrorContext>,
-        /// Entity that has an invalid parent type
-        uid: EntityUID,
-        /// Parent type which was invalid
-        parent_ty: Box<EntityType>, // boxed to avoid this variant being very large (and thus all JsonDeserializationErrors being large)
+        /// Underlying error
+        err: ExtensionFunctionLookupError,
     },
     /// Raised when a JsonValue contains the no longer supported `__expr` escape
     #[error("{0}, invalid escape. The `__expr` escape is no longer supported")]