+++ cedar-policy-core/src/extensions.rs
@@ -40,7 +40,7 @@ lazy_static::lazy_static! {
 /// Holds data on all the Extensions which are active for a given evaluation.
 ///
 /// Clone is cheap for this type.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Copy)]
 pub struct Extensions<'a> {
     /// the actual extensions
     extensions: &'a [Extension],